+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Release Changelog
 -----------------
 
+0.2.0 (2020-11-12)
+~~~~~~~~~~~~~~~~~~
+
+* Refactor instruments into classes with base class (breaking)
+* Rename module: instrumentation -> instruments (breaking)
+* Change args decorator/decorator_kwargs -> instrument/instrument_kwargs (breaking)
+
 0.1.1 (2020-11-11)
 ~~~~~~~~~~~~~~~~~~